@@ -28,39 +28,38 @@
   "main": "lib/index.js",
   "dependencies": {
     "classnames": "latest",
-    "react-highlight-words": "^0.17.0",
-    "react-input-autosize": "^3.0.0",
-    "react-select": "^5.3.2",
-    "react-window": "^1.8.7"
+    "react-highlight-words": "^0.20.0",
+    "react-select": "^5.7.3",
+    "react-window": "^1.8.9"
   },
   "peerDependencies": {
-    "react": "^17.0.0",
-    "react-dom": "^17.0.0"
+    "react": "^17.0.0 || ^18.0.0",
+    "react-dom": "^17.0.0 || ^18.0.0"
   },
   "devDependencies": {
-    "@babel/cli": "^7.14.8",
-    "@babel/core": "^7.14.8",
-    "@babel/plugin-proposal-class-properties": "^7.14.5",
-    "@babel/plugin-proposal-object-rest-spread": "^7.14.7",
-    "@babel/preset-env": "^7.14.9",
-    "@babel/preset-react": "^7.14.5",
-    "@babel/runtime": "^7.14.8",
+    "@babel/cli": "^7.22.5",
+    "@babel/core": "^7.22.5",
+    "@babel/plugin-proposal-class-properties": "^7.18.6",
+    "@babel/plugin-proposal-object-rest-spread": "^7.20.7",
+    "@babel/preset-env": "^7.22.5",
+    "@babel/preset-react": "^7.22.5",
+    "@babel/runtime": "^7.22.5",
     "babel-loader": "^8.2.2",
-    "bootstrap": "^5.1.3",
+    "bootstrap": "^5.3.0",
     "concurrently": "^3.5.1",
-    "css-loader": "^6.7.1",
+    "css-loader": "^6.8.1",
     "html-loader": "^1.3.2",
-    "html-webpack-plugin": "^5.5.0",
-    "husky": "^8.0.0",
+    "html-webpack-plugin": "^5.5.3",
+    "husky": "^8.0.3",
     "prettier": "^2.7.1",
     "pretty-quick": "^3.1.3",
     "prop-types": "^15.8.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "style-loader": "^1.3.0",
-    "webpack": "^5.73.0",
+    "webpack": "^5.87.0",
     "webpack-cli": "^4.10.0",
-    "webpack-dev-server": "^4.9.3"
+    "webpack-dev-server": "^4.15.1"
   },
   "scripts": {
     "dev": "concurrently \"npm run lib:watch\" \"npm start\"",
