@@ -4,21 +4,14 @@ Object.defineProperty(exports, "__esModule", {
   value: true,
 });
 exports.IntelligentTreeSelect = void 0;
-
 var _react = _interopRequireWildcard(require("react"));
-
 var _VirtualizedTreeSelect = require("./VirtualizedTreeSelect");
-
 var _propTypes = _interopRequireDefault(require("prop-types"));
-
 var _Utils = require("./utils/Utils");
-
 var _Constants = _interopRequireDefault(require("./utils/Constants"));
-
 function _interopRequireDefault(obj) {
   return obj && obj.__esModule ? obj : {default: obj};
 }
-
 function _getRequireWildcardCache(nodeInterop) {
   if (typeof WeakMap !== "function") return null;
   var cacheBabelInterop = new WeakMap();
@@ -27,7 +20,6 @@ function _getRequireWildcardCache(nodeInterop) {
     return nodeInterop ? cacheNodeInterop : cacheBabelInterop;
   })(nodeInterop);
 }
-
 function _interopRequireWildcard(obj, nodeInterop) {
   if (!nodeInterop && obj && obj.__esModule) {
     return obj;
@@ -57,7 +49,6 @@ function _interopRequireWildcard(obj, nodeInterop) {
   }
   return newObj;
 }
-
 function _extends() {
   _extends = Object.assign
     ? Object.assign.bind()
@@ -74,7 +65,6 @@ function _extends() {
       };
   return _extends.apply(this, arguments);
 }
-
 class IntelligentTreeSelect extends _react.Component {
   constructor(props, context) {
     super(props, context);
@@ -101,36 +91,27 @@ class IntelligentTreeSelect extends _react.Component {
     };
     this.select = /*#__PURE__*/ _react.default.createRef();
   }
-
   componentDidMount() {
     let data = [];
-
     if (this.props.name && this.props.fetchOptions) {
       data = this._retrieveCachedData();
     }
-
     if (data.length === 0) {
       data = this.props.options;
     }
-
     if (!this.props.simpleTreeData) {
       data = this._simplifyData(this.props.options);
     }
-
     this._addNewOptions(data);
-
     this._loadOptions();
   }
-
   _retrieveCachedData() {
     let cachedData = window.localStorage.getItem(this.props.name);
-
     if (cachedData) {
       cachedData = JSON.parse(cachedData);
       return cachedData.validTo > Date.now() ? cachedData.data : [];
     }
   }
-
   _loadOptions() {
     if (this.state.options.length === 0 && this.props.fetchOptions) {
       if (!this.fetching) {
@@ -138,7 +119,6 @@ class IntelligentTreeSelect extends _react.Component {
       }
     }
   }
-
   _fetchOptions(searchString, optionId, offset, topOption, callback) {
     this.setState({
       isLoadingExternally: true,
@@ -146,61 +126,52 @@ class IntelligentTreeSelect extends _react.Component {
     this.fetching = this._getResponse(searchString, optionId, this.props.fetchLimit, offset, topOption).then(
       (response) => {
         let data;
-
         if (!this.props.simpleTreeData) {
           data = this._simplifyData(response);
         } else {
           data = response;
         }
-
         this.fetching = false;
-
         this._addNewOptions(data);
-
         this.setState({
           isLoadingExternally: false,
         });
-
         if (callback) {
           callback(data);
         }
       }
     );
-  } // If the values are controlled from the outside, it is needed to map them properly to options which Select knows
+  }
 
+  // If the values are controlled from the outside, it is needed to map them properly to options which Select knows
   static getDerivedStateFromProps(props, state) {
     if (!props.valueIsControlled) return state;
-
     if (!props.value) {
       state.passedValue = [];
       state.selectedOptions = [];
       return state;
     }
-
     const values = (0, _Utils.sanitizeArray)(props.value);
     const selectedOpt = (0, _Utils.sanitizeArray)(state.selectedOptions);
     const modifiedPassedValue = [];
     const modifiedSelectedOptions = [];
-
     for (const valueElement of values) {
       const key = valueElement[props.valueKey] ?? valueElement;
       const opt = selectedOpt.find((term) => term[props.valueKey] === key);
-
       if (opt) {
         modifiedSelectedOptions.push(opt);
       } else {
         modifiedPassedValue.push(key);
       }
     }
-
     state.passedValue = modifiedPassedValue;
     state.selectedOptions = modifiedSelectedOptions;
     return state;
   }
+
   /**
    * Resets the option, forcing the component to reload them from the server/reload them from props.
    */
-
   resetOptions() {
     if (this.props.fetchOptions) {
       this.setState(
@@ -224,32 +195,31 @@ class IntelligentTreeSelect extends _react.Component {
       );
     }
   }
+
   /**
    * Focuses the select input.
    */
-
   focus() {
     if (this.select.current) {
       this.select.current.focus();
     }
   }
+
   /**
    * Blurs the select input.
    */
-
   blurInput() {
     if (this.select.current) {
       this.select.current.blurInput();
     }
   }
+
   /**
    * Gets the current options provided by this component.
    */
-
   getOptions() {
     return this.state.options.slice();
   }
-
   componentDidUpdate(prevProps) {
     if (!this.props.fetchOptions && prevProps.options !== this.props.options) {
       this.setState(
@@ -263,10 +233,8 @@ class IntelligentTreeSelect extends _react.Component {
       );
     }
   }
-
   _isInHistory(searchString) {
     searchString = searchString.toLowerCase();
-
     for (let i = 0; i < this.history.length; i++) {
       if (this.history[i].searchString.toLowerCase() === searchString) {
         if (Date.now() < this.history[i].validTo) {
@@ -274,30 +242,24 @@ class IntelligentTreeSelect extends _react.Component {
         }
       }
     }
-
     return false;
   }
-
   _simplifyData(responseData) {
     let result = [];
     const {valueKey, childrenKey} = this.props;
     if (!responseData || responseData.length === 0) return result;
-
     for (let i = 0; i < responseData.length; i++) {
       //deep clone
       let data = JSON.parse(JSON.stringify(responseData[i]));
       result = result.concat(data);
       const childrenArr = (0, _Utils.sanitizeArray)(data[childrenKey]);
-
       if (childrenArr.length > 0) {
         result = result.concat(this._simplifyData(childrenArr));
         data[childrenKey] = childrenArr.map((xdata) => xdata[valueKey]);
       }
     }
-
     return result;
   }
-
   _parseOptionLifetime(value) {
     let optionLifetime = {
       days: 0,
@@ -305,7 +267,6 @@ class IntelligentTreeSelect extends _react.Component {
       minutes: 30,
       seconds: 0,
     };
-
     if (/^(\d+d)?(\d+h)?(\d+m)?(\d+s)?$/.test(value)) {
       let tmp = /^(\d+d)?(\d+h)?(\d+m)?(\d+s)?$/.exec(value);
       optionLifetime = {
@@ -317,10 +278,8 @@ class IntelligentTreeSelect extends _react.Component {
     } else {
       throw new Error("Invalid optionLifetime. Expecting format: e.g. 1d10h5m6s ");
     }
-
     return optionLifetime;
   }
-
   _getValidForInSec(optionLifetime) {
     optionLifetime = this._parseOptionLifetime(optionLifetime);
     let res = 0;
@@ -330,7 +289,6 @@ class IntelligentTreeSelect extends _react.Component {
     res += isNaN(optionLifetime.days) ? 0 : optionLifetime.days * 60 * 60 * 24;
     return res * 1000;
   }
-
   _getRootNodesCount() {
     let count = 0;
     this.state.options.forEach((option) => {
@@ -338,7 +296,6 @@ class IntelligentTreeSelect extends _react.Component {
     });
     return count;
   }
-
   async _getResponse(searchString, optionID, limit, offset, option) {
     return this.props.fetchOptions
       ? await this.props.fetchOptions({
@@ -350,28 +307,23 @@ class IntelligentTreeSelect extends _react.Component {
         })
       : [];
   }
-
   _onInputChange(searchString) {
     if (searchString && this.props.fetchOptions) {
       let dataCached = false;
-
       for (let i = searchString.length; i > 0; i--) {
         if (dataCached) break;
         let substring = searchString.substring(0, i);
         dataCached = this._isInHistory(substring);
       }
-
       if (!dataCached && !this.fetching) {
         let data = [];
         let offset = 0;
         this.state.options.forEach((option) => {
           if (option.depth === 0) offset++;
         });
-
         if (this.searchTimer) {
           clearTimeout(this.searchTimer);
         }
-
         if (this.props.searchDelay) {
           this.searchTimer = setTimeout(() => {
             this._invokeSearch(searchString, offset);
@@ -381,30 +333,23 @@ class IntelligentTreeSelect extends _react.Component {
         }
       }
     }
-
     this.searchString = searchString;
-
     if ("onInputChange" in this.props) {
       this.props.onInputChange(searchString);
     }
   }
-
   _invokeSearch(searchString, offset) {
     this._fetchOptions(searchString, "", offset, undefined, () => {
       this._addToHistory(searchString, Date.now() + this._getValidForInSec(this.props.optionLifetime));
-
       this.select.current.filterValues(searchString);
     });
   }
-
   _onScroll(data) {
     const {clientHeight, scrollHeight, scrollTop} = data;
-
     if (scrollHeight - scrollTop <= 2.5 * clientHeight && !this.fetching && this.props.fetchOptions) {
       // this.fetching = true;
       let totalOptionsHeight = 0;
       let topOptionIndex = 0;
-
       for (topOptionIndex; topOptionIndex < this.state.options.length; topOptionIndex++) {
         const option = this.state.options[topOptionIndex];
         totalOptionsHeight +=
@@ -413,17 +358,14 @@ class IntelligentTreeSelect extends _react.Component {
                 option,
               })
             : this.props.optionHeight;
-
         if (totalOptionsHeight >= scrollTop) {
           break;
         }
       }
-
       const topOption = this.state.options[topOptionIndex];
       let parentOption = this.state.options.find((option) => option[this.props.valueKey] === topOption.parent);
       let offset = parentOption ? parentOption[this.props.childrenKey].length : this._getRootNodesCount();
       let parentOptionValue = parentOption ? parentOption[this.props.valueKey] : "root";
-
       if (!this.completedNodes[parentOptionValue]) {
         //fetch child options that are not completed
         this._fetchOptions("", topOption.parent, offset, topOption, (fetchedData) => {
@@ -435,57 +377,46 @@ class IntelligentTreeSelect extends _react.Component {
       }
     }
   }
-
   _onOptionToggle(option) {
     if (!option.expanded) {
       let dataCached = this._isInHistory(option[this.props.valueKey]);
-
       if (!dataCached) {
         this.setState({
           isLoadingExternally: true,
         });
         option.fetchingChild = true;
         let data = [];
-
         this._getResponse("", option[this.props.valueKey], this.props.fetchLimit, 0, option).then((response) => {
           if (!this.props.simpleTreeData) {
             data = this._simplifyData(response);
           } else {
             data = response;
           }
-
           if (data.length < this.props.fetchLimit) {
             this.completedNodes[option[this.props.valueKey]] = true;
           }
-
           this._addToHistory(
             option[this.props.valueKey],
             Date.now() + this._getValidForInSec(this.props.optionLifetime)
           );
-
           delete option.fetchingChild;
-
           if (data.length > 0) {
             this._addNewOptions(data);
           }
-
           this.setState({
             isLoadingExternally: false,
           });
         });
       }
     }
-
     this.forceUpdate();
   }
-
   _valueRenderer({children, data}) {
     if (this.props.valueRenderer) {
       // On initial render, there can be empty options
       if (!children) return null;
       return this.props.valueRenderer(children, data);
     }
-
     const {valueKey, labelKey, getOptionLabel} = this.props;
     const value = data[valueKey];
     if ((0, _Utils.isURL)(value))
@@ -502,30 +433,24 @@ class IntelligentTreeSelect extends _react.Component {
       );
     return children;
   }
-
   _onOptionCreate(option) {
     if (option.parent) {
       this._addChildrenToParent(option[this.props.valueKey], option.parent);
     }
-
     this._addNewOptions([option]);
-
     if ("onOptionCreate" in this.props) {
       this.props.onOptionCreate(option);
     }
   }
-
   _addNewOptions(newOptions) {
     const {childrenKey, fetchOptions, name, optionLifetime} = this.props;
     let mergedArr;
-
     if (this.state.options.length === 0) {
       newOptions.forEach((no) => (no[childrenKey] = (0, _Utils.sanitizeArray)(no[childrenKey])));
       mergedArr = newOptions;
     } else {
       mergedArr = this._mergeOptionArrays(this.state.options, newOptions);
     }
-
     if (name && fetchOptions) {
       window.localStorage.setItem(
         name,
@@ -535,22 +460,20 @@ class IntelligentTreeSelect extends _react.Component {
         })
       );
     }
-
     if (newOptions.length > 0) {
       this._finalizeSelectedOptions(newOptions, mergedArr);
     }
-
     this.setState({
       options: mergedArr,
       update: ++this.state.update,
     });
   }
-
   _mergeOptionArrays(originalOptions, newOptions) {
     const {valueKey, childrenKey} = this.props;
     let options = originalOptions.concat(newOptions);
-    let mergedArr = []; //merge options
+    let mergedArr = [];
 
+    //merge options
     while (options.length > 0) {
       let currOption = options.shift();
       currOption[childrenKey] = (0, _Utils.sanitizeArray)(currOption[childrenKey]);
@@ -567,15 +490,14 @@ class IntelligentTreeSelect extends _react.Component {
       mergedArr.push((0, _Utils.monotonicAssign)({}, currOption, ...conflicts.reverse()));
       options = optionsToReplace;
     }
-
     return mergedArr;
-  } //Check if new options contain selected value
+  }
 
+  //Check if new options contain selected value
   _finalizeSelectedOptions(addedOptions, parsedOptions) {
     const foundOptions = [];
     let previouslySelected = (0, _Utils.sanitizeArray)(this.state.passedValue);
     let newSelected = (0, _Utils.sanitizeArray)(this.state.selectedOptions);
-
     for (const selectedOpt of previouslySelected) {
       const key = selectedOpt[this.props.valueKey] ?? selectedOpt;
       const option = addedOptions.find((term) => term[this.props.valueKey] === key);
@@ -584,47 +506,40 @@ class IntelligentTreeSelect extends _react.Component {
       const optionParsed = parsedOptions.find((term) => term[this.props.valueKey] === key);
       newSelected.push(optionParsed);
     }
+    this._addSelectedOption(newSelected);
 
-    this._addSelectedOption(newSelected); //remove already found options
-
+    //remove already found options
     for (const foundOption of foundOptions) {
       previouslySelected = previouslySelected.filter((term) => {
         return term !== foundOption;
       });
     }
-
     this.setState({
       passedValue: previouslySelected,
     });
   }
-
   _addChildrenToParent(childrenID, parentID) {
     let parentOption = this.state.options.find((x) => x[this.props.valueKey] === parentID);
     let children = parentOption[this.props.childrenKey];
     if (children.indexOf(childrenID) === -1) children.push(childrenID);
   }
-
   _onChange(options) {
     this._addSelectedOption(options);
-
     if (this.props.onChange) {
       this.props.onChange(options);
     }
   }
-
   _addSelectedOption(selectedOptions) {
     this.setState({
       selectedOptions,
     });
   }
-
   _addToHistory(searchString, validTo) {
     this.history.unshift({
       searchString,
       validTo,
     });
   }
-
   render() {
     let listProps = {};
     listProps.onScroll = this.props.onScroll || this._onScroll;
@@ -667,7 +582,6 @@ class IntelligentTreeSelect extends _react.Component {
     );
   }
 }
-
 exports.IntelligentTreeSelect = IntelligentTreeSelect;
 IntelligentTreeSelect.propTypes = {
   autoFocus: _propTypes.default.bool,
