@@ -4,25 +4,16 @@ Object.defineProperty(exports, "__esModule", {
   value: true,
 });
 exports.VirtualizedTreeSelect = void 0;
-
 var _react = _interopRequireWildcard(require("react"));
-
 var _reactSelect = _interopRequireWildcard(require("react-select"));
-
 var _propTypes = _interopRequireDefault(require("prop-types"));
-
 var _Option = _interopRequireDefault(require("./Option"));
-
 var _Constants = _interopRequireDefault(require("./utils/Constants"));
-
 var _reactWindow = require("react-window");
-
 var _Utils = require("./utils/Utils");
-
 function _interopRequireDefault(obj) {
   return obj && obj.__esModule ? obj : {default: obj};
 }
-
 function _getRequireWildcardCache(nodeInterop) {
   if (typeof WeakMap !== "function") return null;
   var cacheBabelInterop = new WeakMap();
@@ -31,7 +22,6 @@ function _getRequireWildcardCache(nodeInterop) {
     return nodeInterop ? cacheNodeInterop : cacheBabelInterop;
   })(nodeInterop);
 }
-
 function _interopRequireWildcard(obj, nodeInterop) {
   if (!nodeInterop && obj && obj.__esModule) {
     return obj;
@@ -61,7 +51,6 @@ function _interopRequireWildcard(obj, nodeInterop) {
   }
   return newObj;
 }
-
 function ownKeys(object, enumerableOnly) {
   var keys = Object.keys(object);
   if (Object.getOwnPropertySymbols) {
@@ -74,7 +63,6 @@ function ownKeys(object, enumerableOnly) {
   }
   return keys;
 }
-
 function _objectSpread(target) {
   for (var i = 1; i < arguments.length; i++) {
     var source = null != arguments[i] ? arguments[i] : {};
@@ -90,8 +78,8 @@ function _objectSpread(target) {
   }
   return target;
 }
-
 function _defineProperty(obj, key, value) {
+  key = _toPropertyKey(key);
   if (key in obj) {
     Object.defineProperty(obj, key, {value: value, enumerable: true, configurable: true, writable: true});
   } else {
@@ -99,7 +87,20 @@ function _defineProperty(obj, key, value) {
   }
   return obj;
 }
-
+function _toPropertyKey(arg) {
+  var key = _toPrimitive(arg, "string");
+  return typeof key === "symbol" ? key : String(key);
+}
+function _toPrimitive(input, hint) {
+  if (typeof input !== "object" || input === null) return input;
+  var prim = input[Symbol.toPrimitive];
+  if (prim !== undefined) {
+    var res = prim.call(input, hint || "default");
+    if (typeof res !== "object") return res;
+    throw new TypeError("@@toPrimitive must return a primitive value.");
+  }
+  return (hint === "string" ? String : Number)(input);
+}
 function _extends() {
   _extends = Object.assign
     ? Object.assign.bind()
@@ -116,7 +117,6 @@ function _extends() {
       };
   return _extends.apply(this, arguments);
 }
-
 class VirtualizedTreeSelect extends _react.Component {
   constructor(props, context) {
     super(props, context);
@@ -142,135 +142,116 @@ class VirtualizedTreeSelect extends _react.Component {
     };
     this.select = /*#__PURE__*/ _react.default.createRef();
   }
-
   componentDidMount() {
     this._processOptions();
   }
-
   componentDidUpdate(prevProps) {
     if (this.props.update > prevProps.update) {
       this._processOptions();
     }
   }
-
   focus() {
     this.select.current.focus();
   }
-
   blurInput() {
     if (this.select.current) {
       this.select.current.blur();
     }
   }
-
   _processOptions() {
     this.data = {};
     const keys = [];
     this.props.options.forEach((option) => {
-      const optionID = option[this.props.valueKey]; // Value property is needed for correct rendering of selected options
-
+      const optionID = option[this.props.valueKey];
+      // Value property is needed for correct rendering of selected options
       option.value = optionID;
       this.data[optionID] = option;
       keys.push(optionID);
-    }); // Utilize the fact that set has stable iteration order (~ insertion order)
-
+    });
+    // Utilize the fact that set has stable iteration order (~ insertion order)
     const sortedArr = new Set();
     keys.forEach((key) => {
       let option = this.data[key];
-
       if (!option.parent) {
         this._calculateDepth(key, 0, null, new Set(), sortedArr);
       }
     });
-    let options = [...sortedArr]; // Expands the whole tree on the initial render
+    let options = [...sortedArr];
 
+    // Expands the whole tree on the initial render
     if (this.props.expanded && !this.state.initialExpansion && options.length > 0) {
       for (const option of options) {
         this.toggledOptions.push(option);
         option.expanded = true;
       }
-
       this.setState({
         initialExpansion: true,
       });
     }
-
     this.setState({
       options,
     });
   }
-
   _findOption(dataset, searchedOption) {
     if (!searchedOption || !dataset) return null;
     let options = dataset.filter((el) => el[this.props.valueKey] === searchedOption[this.props.valueKey]);
-
     for (const option of options) {
       if ((0, _Utils.arraysAreEqual)(option.path, searchedOption.path)) {
         return option;
       }
     }
-
     return null;
   }
-
   _findOptionWithParent(dataset, searchedOptionKey, parent) {
     let options = dataset.filter((el) => el[this.props.valueKey] === searchedOptionKey);
     return options.find((el) => el?.parent === parent);
   }
-
   _calculateDepth(key, depth, parent, visited, sortedArr) {
     let option = this.data[key];
-
     if (!option || visited.has(key)) {
       return;
-    } //Checks whether the array of items already contain an option with the same valueKey (ID)
-
+    }
+    //Checks whether the array of items already contain an option with the same valueKey (ID)
     if (sortedArr.has(option)) {
       //Deep copy of option, needed to distinguish option for multiple subtrees
       option = structuredClone(option);
     }
-
     sortedArr.add(option);
-    visited.add(key); //Sets the idempotent properties
+    visited.add(key);
 
+    //Sets the idempotent properties
     option.depth = depth;
     option.parent = parent;
     option.path = [...visited];
-    option.expanded = false; //It can happen that the option is already loaded in the state
+    option.expanded = false;
+
+    //It can happen that the option is already loaded in the state
     //If so, set the correct expanded value from the state options
     //It is needed to check its full path to determine whether it is the correct option
-
     let existingOption = this._findOption(this.state.options, option);
-
     if (existingOption) {
       option.expanded = existingOption.expanded;
     }
-
     option[this.props.childrenKey].forEach((childID) => {
       this._calculateDepth(childID, depth + 1, option, visited, sortedArr);
     });
   }
-
   filterOption(candidate, inputValue) {
     const option = candidate.data;
     inputValue = inputValue.trim().toLowerCase();
-
     if (inputValue.length === 0) {
       return !option.parent || option.parent?.expanded;
     } else {
       return option.visible;
     }
   }
-
   filterValues(searchInput) {
     // when the fetch is delayed, it can cause incorrect filter render, this prevents it from happening
     if (this.select.current.inputRef.value !== searchInput) {
       searchInput = this.select.current.inputRef.value;
     }
-
     if (searchInput === "") return;
     const matches = [];
-
     for (let option of this.state.options) {
       if (this.matchCheck(searchInput, (0, _Utils.getLabel)(option, this.props.labelKey, this.props.getOptionLabel))) {
         option.visible = true;
@@ -279,7 +260,6 @@ class VirtualizedTreeSelect extends _react.Component {
         option.visible = false;
       }
     }
-
     for (let match of matches) {
       while (match.parent !== null) {
         match = match.parent;
@@ -287,113 +267,92 @@ class VirtualizedTreeSelect extends _react.Component {
         match.visible = true;
       }
     }
-
     this.forceUpdate();
   }
-
   matchCheckFull(searchInput, optionLabel) {
     return optionLabel.toLowerCase().indexOf(searchInput.toLowerCase()) !== -1;
   }
-
   _onInputChange(input) {
     // Make the expensive calculation only when input has been really changed
     if (this.searchString !== input && input.length !== 0) {
       this.filterValues(input);
     }
-
     this.searchString = input;
-
     if ("onInputChange" in this.props) {
       this.props.onInputChange(input);
-    } // Collapses items which were expanded by the search
-
+    }
+    // Collapses items which were expanded by the search
     if (input.length === 0) {
       for (let option of this.state.options) {
         option.expanded = !!this._findOption(this.toggledOptions, option);
       }
     }
   }
-
   _removeChildrenFromToggled(option) {
     if (option === undefined) return;
-
     for (const subTermId of option[this.props.childrenKey]) {
       const subTerm = this._findOptionWithParent(this.state.options, subTermId, option);
-
       const toggledItem = this._findOption(this.toggledOptions, subTerm);
-
       this.toggledOptions = this.toggledOptions.filter((term) => term !== toggledItem);
-
       this._removeChildrenFromToggled(subTerm);
     }
   }
-
   _onOptionClose(option) {
     if (option === undefined) return;
     option.expanded = false;
-
     for (const subTermId of option[this.props.childrenKey]) {
       const subTerm = this._findOptionWithParent(this.state.options, subTermId, option);
-
       this._onOptionClose(subTerm);
     }
   }
-
   _onOptionToggle(option) {
     // disables option expansion/collapsion when search string is present
     if (this.searchString !== "") {
       return;
     }
-
     this.props.onOptionToggle(option);
-
     if (option.expanded) {
       this._onOptionClose(option);
     } else {
       option.expanded = true;
-    } // Adds/removes references for toggled items
+    }
 
+    // Adds/removes references for toggled items
     if (option.expanded) {
       this.toggledOptions.push(option);
     } else {
       const toggledItem = this._findOption(this.toggledOptions, option);
-
       this.toggledOptions = this.toggledOptions.filter((el) => el !== toggledItem);
-
       this._removeChildrenFromToggled(option);
     }
-  } //When selecting an option, we want to ensure that the path to it is expanded
-  //Path is saved in toggledOptions
+  }
 
+  //When selecting an option, we want to ensure that the path to it is expanded
+  //Path is saved in toggledOptions
   _onOptionSelect(props) {
     props.selectOption(props.data);
     const isSelected = props.isSelected;
     if (isSelected) return;
     let parent = props.data.parent;
-
     while (parent) {
       let option = this._findOption(this.toggledOptions, parent);
-
       if (!option) {
         parent.expanded = true;
         this.toggledOptions.push(parent);
       }
-
       parent = option?.parent ?? parent.parent;
     }
-  } //When using custom option, it is needed to set focusedOption manually
+  }
 
+  //When using custom option, it is needed to set focusedOption manually
   _onOptionHover(option) {
     this.select.current.setState({
       focusedOption: option,
     });
   }
-
   render() {
     const props = this.props;
-
     const styles = this._prepareStyles();
-
     const filterOptions = props.filterOptions || this.filterOption;
     const optionRenderer = this.props.optionRenderer || _Option.default;
     return /*#__PURE__*/ _react.default.createElement(
@@ -428,7 +387,6 @@ class VirtualizedTreeSelect extends _react.Component {
       )
     );
   }
-
   _prepareStyles() {
     return {
       dropdownIndicator: (provided, state) =>
@@ -507,10 +465,10 @@ class VirtualizedTreeSelect extends _react.Component {
         ),
     };
   }
-} // Wrapper for MenuList, it doesn't do anything, it is only needed for correct pass of the onScroll prop
+}
 
+// Wrapper for MenuList, it doesn't do anything, it is only needed for correct pass of the onScroll prop
 exports.VirtualizedTreeSelect = VirtualizedTreeSelect;
-
 const Menu = (props) => {
   return /*#__PURE__*/ _react.default.createElement(
     _reactSelect.components.Menu,
@@ -527,7 +485,6 @@ const Menu = (props) => {
             if (event.button !== 0) {
               return;
             }
-
             event.stopPropagation();
             event.preventDefault();
           },
@@ -536,15 +493,16 @@ const Menu = (props) => {
     }),
     props.children
   );
-}; // Component for efficient rendering
+};
 
+// Component for efficient rendering
 const MenuList = (props) => {
   const {children} = props;
-  const {optionHeight, maxHeight} = props.selectProps; // We need to check whether the passed object contains items or loading/empty message
+  const {optionHeight, maxHeight} = props.selectProps;
 
+  // We need to check whether the passed object contains items or loading/empty message
   let values;
   let height;
-
   if (Array.isArray(children)) {
     values = children;
     height = Math.min(maxHeight, optionHeight * values.length);
@@ -559,7 +517,6 @@ const MenuList = (props) => {
     ];
     height = 40;
   }
-
   return /*#__PURE__*/ _react.default.createElement(
     _reactWindow.FixedSizeList,
     {
@@ -578,7 +535,6 @@ const MenuList = (props) => {
       )
   );
 };
-
 VirtualizedTreeSelect.propTypes = {
   autoFocus: _propTypes.default.bool,
   childrenKey: _propTypes.default.string,
