@@ -1,35 +1,32 @@
 "use strict";
 
 Object.defineProperty(exports, "__esModule", {
-  value: true
+  value: true,
 });
 Object.defineProperty(exports, "IntelligentTreeSelect", {
   enumerable: true,
   get: function () {
     return _IntelligentTreeSelect.IntelligentTreeSelect;
-  }
+  },
 });
 Object.defineProperty(exports, "ToggleMinusIcon", {
   enumerable: true,
   get: function () {
     return _Icons.ToggleMinusIcon;
-  }
+  },
 });
 Object.defineProperty(exports, "TogglePlusIcon", {
   enumerable: true,
   get: function () {
     return _Icons.TogglePlusIcon;
-  }
+  },
 });
 Object.defineProperty(exports, "VirtualizedTreeSelect", {
   enumerable: true,
   get: function () {
     return _VirtualizedTreeSelect.VirtualizedTreeSelect;
-  }
+  },
 });
-
 var _IntelligentTreeSelect = require("../lib/components/IntelligentTreeSelect");
-
 var _VirtualizedTreeSelect = require("../lib/components/VirtualizedTreeSelect");
-
 var _Icons = require("../lib/components/Icons");
